replace simpleini with git submodule or system lib if found
