+++ .gitmodules
@@ -16,3 +16,6 @@
 [submodule "3rdparty/zlib"]
 	path = 3rdparty/zlib
 	url = https://github.com/madler/zlib.git
+[submodule "3rdparty/simpleini"]
+	path = 3rdparty/simpleini
+	url = https://github.com/brofield/simpleini.git
