+++ libs/config/CMakeLists.txt
@@ -2,8 +2,10 @@
 list(APPEND config_EXTRA_SRCS		"${MRPT_LIBS_ROOT}/config/*.cpp" "${MRPT_LIBS_ROOT}/config/*.h")
 list(APPEND config_EXTRA_SRCS_NAME 	"config" "config")
 
-list(APPEND config_EXTRA_SRCS       "${MRPT_LIBS_ROOT}/config/src/simpleini/*.h")
-list(APPEND config_EXTRA_SRCS_NAME  "config/simpleini")
+if (NOT CMAKE_MRPT_HAS_SIMPLEINI_SYSTEM)
+	list(APPEND config_EXTRA_SRCS       "${MRPT_LIBS_ROOT}/config/src/simpleini/*.h")
+	list(APPEND config_EXTRA_SRCS_NAME  "config/simpleini")
+endif()
 
 define_mrpt_lib(
 	config 		# Lib name
@@ -14,5 +16,6 @@ define_mrpt_lib(
 	)
 
 if(BUILD_mrpt-config)
+	target_include_directories(config PRIVATE "${SIMPLEINI_INCLUDE_DIRS}")
 
 endif()
