+++ parse-files/config.h.in
@@ -94,6 +94,9 @@ ${CMAKE_MRPT_BUILD_SHARED_LIB}
 /** jsoncpp is present */
 #define MRPT_HAS_JSONCPP ${CMAKE_MRPT_HAS_JSONCPP}
 
+#define MRPT_HAS_SIMPLEINI ${CMAKE_MRPT_HAS_SIMPLEINI}
+#define MRPT_HAS_SIMPLEINI_SYSTEM ${CMAKE_MRPT_HAS_SIMPLEINI_SYSEM}
+
 /** These two values are detected in Eigen when building MRPT, so we have
   the same settings given the user-provided flags  */
 #define MRPT_MAX_ALIGN_BYTES           ${EIGEN_MAX_ALIGN_BYTES}
